vector.mo: remove fields related to superblocks from vector
The values are implicit and can be computed on the fly when needed.
Doing so saves us constant memory overhead which matters for small
vectors. Three values are removed from the structure.

As a side effect, it is also easy to initialize the empty
vector with 0 data blocks. This saves more memory for empty
vectors.

This commit is not complete. Size reduction is not done yet.