Fix protocol with new logic
There was the following bug: Suppose two chunks in a row produce an
error and the errors come back in order. The first error will cause
a rewind and sending chunks will resume right away. Both chunks will
get sent again. When the second error comes back then we cannot tell
whether it corresponds to the first time the batch was sent or the
second time it was sent. The index/headIndex values are equal. So
the second sending may be successful but we treat it as an error
because we see the error from the first time it was sent.

The problem is that we resumed sending too early. The new logic
waits after any error for _all_ responses to come before resuming.