Remove #NotRegistered error
SharedStream is better self-contained. It is better if the library
consisting of StreamSender/Receiver has to handle only those errors
that one of its own classes issues.

Moreover, in the application, for reasons to protect against spam
calls, we should not respond to a caller who isn't registered. We
should just trap. Therefore the error #NotRegistered should not
exist.