Remove #BrokenPipe error
The value delivered in #BrokenPipe is not used anymore in the new logic.
Hence we remove the error because that simplifies the code (one case
less in the sender logic). Instead, the receiver throws (CANISTER_REJECT)
in case of a broken pipe. This is handled by the sender in the same way
as a failed delivery.