Send StreamClosed only if pipe not broken
We now test for broken pipe before we check if the stream is closed.
Then there is only one response with a StreamClosed error, not multiple.
Only the first gapless batch after the stream's closing position will
get this error. The later batches will throw like any other broken pipe
situation.

The position value (length of the closed stream) is then superfluous.
Since StreamClosed was the last remaining case in the ChunkError varint
we get rid of the variant entirely. Error handling code in the sender
becomes simpler.