StreamReceiver: update lastChunkReceived only if we make progress
We update lastChunkReceived if a chunk is actually processed, not in
case of a broken pipe. In other words we only update it if the protocol
makes progress. This behavior is more meaningful. Otherwise, if only
a fraction of chunks make it through then it could happen that we never
timeout despite the protocol making zero progress. This is not the user
experience we want for which the timeout was invented.