SharedStreams: remove dependency on QueueBuffer
Instead of importing QueueBuffer we define a single object directly
inside the StreamSender. The benefits:
- less code overall because we only put inside the object what we need
- we can also add the chunk() function to the same object

Note: We currently expose the buffer inside the queue object and let
calling code access it directly as queue.buf. This could be better
encapsulated if desired. But we leave that for a subsequent commit.