Test bug (#130)
* Test that exposes bug on fee change recalculation

* fix: skip locked principals in recalculation

* Remove consolidation retry after BadFee

It isn't necessary because the next trigger will take care of it
anyway.

* introduce setNewFee function

* Count the number of tree accesses and report them in tests

This code will be removed again later. While we have the code we
can work on improving the number of accesses required.

* Export journal length

* Add new test approach

With this approach we have flexibility to simulate ledger
behaviour. We can also simulate call delivery failures such as
canister_reject, etc.

* Exposing bug in depositDelta for dust

With the new bookkeeping we don't store dust deposits anymore at
all. This means we have to report a dust deposit with depositDelta
= 0. Otherwise a dust deposit will be reported with over and over
again (as non-zero) by subsequent calls to notify().

* fix notify return value for dust

* reorder lines in notify()

* Add test for backlog recalculation

* Introduce async tests, expose bug

We introduce a test that can simulate concurrency. We can hold
back a response and control when it returns. For example we
increase the fee while a notify call is under way.

* Add another concurrency test, adjust MockLedger

We have to adjust MockLedger for this and the previous test.
Holding back a response requires an async statement which is not
allowed in query calls. Therefore we change icrc1_balance_of from
query to non-query.

* Fix first test (scenario 1)

* Enable scenario 2 test as well

---------

Co-authored-by: Denys <154005444+reginleif888@users.noreply.github.com>