recalculateDepositRegistry and notify
We realized that recalculateDepositRegistry andenotify can overlap freely.  We
distinguish the locks coming from notify from the ones from consolidate.  Only
if p is locked from consolidation then we skip over p in
recalculateDepositRegistry.

That way, at the end of notify we don't have to reprocess p for a potential fee
change anymore. Because that has already been done. This simplifies code and
debits the user earlier.