Initial commit for a new test approach
* mockLedger is now an object (not actor)
* can handle arbitrary number of simultaneous calls  to the same function
* can simulate asynchronous errors (canister_reject)
* can reorder responses

Required small changes to the TokenHandler class itself, but there are no
obvious downsides to those changes.

Retrofitted old test script async.test.mo to the changes.

Needed to remove `query` declaration from the ICRC1 interface, but there's no
obvious downside to doing that.

Provide a short new script that uses the new functionality as an example. It is
not yet a meaningful test.

Small nuisance: need one empty await-async as a delay statement in the new
script. It does not seem possible to avoid. It is rooted in the fact that the
ICRC1 ledger api that is passed into TokenHandler is async and not async*.
There needs to be delay so that the call made by TokenHandler can read the
register before it gets overwritten by the preparation of the next call.