Refactor consolidation
When consolidation start we take the value out of the deposit registry, i.e. we
set the value to 0 (but keep the lock).  When a fee change happens during
consolidation then it automatically skips this entry because the value is 0.
In recalculateDepositRegistry we no longer have to pay attention to locks.
When consolidation comes back successful we don't have to do anything.  When
consolidation comes back with an error then we have to debit the user and put
the value back into the deposit registry.