Refactor consolidation (#142)
Refactor consolidation:

When consolidation starts we now take the value out of the deposit registry, i.e. we
set the value to 0 (but keep the lock).  When a fee change happens during
consolidation then it automatically skips this entry because the value is 0.
In recalculateDepositRegistry we no longer have to pay attention to locks.
When consolidation comes back successful we don't have to do anything.  When
consolidation comes back with an error then we have to debit the user and put
the value back into the deposit registry.

This allows simplifications:
* remove multiple lock types, make lock a Bool
* recalculation does not care about locks anymore
* recalculation code is more modular, simpler
* can move operations inside NatMap (minimum, iterate)

This cleanly separates recalculation into two steps:
- applying the new fee as a minimum, erasing all deposits that are below it
- adjusting credits that are in the backlog by the fee difference between old
  and new fee.

Other changes:
* in notify(), trigger only when increment detected
* new NatMap function obtainAnyLock()
* reduce code repetition with clean() function
* rename updateFee() to fetchFee()
* replace info() function by getCredit() and trackedDeposit()
* update CI

The new trackedDeposit() function returns null if the principal is locked which
is the desired behaviour because while notification or consolidation are under
way the returned value could be too low or too high.

The PR saves tree lookups, now down from
51 to 42 in tests.