Nat map (#139)
* Initial commit for a new test approach

* mockLedger is now an object (not actor)
* can handle arbitrary number of simultaneous calls  to the same function
* can simulate asynchronous errors (canister_reject)
* can reorder responses

Required small changes to the TokenHandler class itself, but there are no
obvious downsides to those changes.

Retrofitted old test script async.test.mo to the changes.

Needed to remove `query` declaration from the ICRC1 interface, but there's no
obvious downside to doing that.

Provide a short new script that uses the new functionality as an example. It is
not yet a meaningful test.

Small nuisance: need one empty await-async as a delay statement in the new
script. It does not seem possible to avoid. It is rooted in the fact that the
ICRC1 ledger api that is passed into TokenHandler is async and not async*.
There needs to be delay so that the call made by TokenHandler can read the
register before it gets overwritten by the preparation of the next call.

* Add forgotten file, example how to use new approach

* Remove old tests

fee_change.test.mo
token_handler.test.mo
MockLedger.mo

* remove debug output

* Streamline boilerplate code in async2.test.mo

* Remove ledger principal from TokenHandler constructor

* fix test

* introduce clear() function, improve example

* Make default case(_) more explicit

* clean positioning of underwayFunds accounting

* expose bug

add a test for when a consolidation transfer returns an async error

* fix bug

* move queuedFunds tracking one level up

Do += and -= on the same level, in trigger()

* revert previous commit (didn't work)

* Replace DepositRegistry with NatMapWithLock, first step

* Replace depositedFunds_ counter with internal one from NatMap

* Replace all uses of updateDeposit(), reduce tree lookups

Tree lookups in a test are down from 130 to 57

* Return value from `obtainLock`

Tree lookups now at 51 (down from 130 originally)

* Remove set() function from NatMap, no longer used

* remove old comment line

* Start structuring tests

* Add test for decreasing fee, fix bug

* Rewrite tests related to notify using old approach

* Expand with consolidation tests

* Withdraw tests

* Reset mock ledger state

* Small variable naming fix

* Add test for fee change while locked 0-deposit exists, fix related bug

* Expose and fix bug with recalculation on fee change

* assertIntegrity public

* complete merge

* refactor recalculateDepositRegistry

* Remove queuedFunds variable

queuedFunds is depositedFunds minus underwayFunds
we no longer need to track it

* State function (#141)

* add state() function returning record
* use state() function in tests
* remove multiple getter functions that are now in state()
* Move release() call to inside consolidate() function
* Change state() assertion in test to expose the line number

* Refactor consolidation (#142)

Refactor consolidation:

When consolidation starts we now take the value out of the deposit registry, i.e. we
set the value to 0 (but keep the lock).  When a fee change happens during
consolidation then it automatically skips this entry because the value is 0.
In recalculateDepositRegistry we no longer have to pay attention to locks.
When consolidation comes back successful we don't have to do anything.  When
consolidation comes back with an error then we have to debit the user and put
the value back into the deposit registry.

This allows simplifications:
* remove multiple lock types, make lock a Bool
* recalculation does not care about locks anymore
* recalculation code is more modular, simpler
* can move operations inside NatMap (minimum, iterate)

This cleanly separates recalculation into two steps:
- applying the new fee as a minimum, erasing all deposits that are below it
- adjusting credits that are in the backlog by the fee difference between old
  and new fee.

Other changes:
* in notify(), trigger only when increment detected
* new NatMap function obtainAnyLock()
* reduce code repetition with clean() function
* rename updateFee() to fetchFee()
* replace info() function by getCredit() and trackedDeposit()
* update CI

The new trackedDeposit() function returns null if the principal is locked which
is the desired behaviour because while notification or consolidation are under
way the returned value could be too low or too high.

The PR saves tree lookups, now down from
51 to 42 in tests.

* rename obtainAnyLock() to nextLock(), simplify obtainLock()

* remove unused NatMap.entries() function

* Simplify consolidation transfer result handling

---------

Co-authored-by: Denys Kushnarov <154005444+reginleif888@users.noreply.github.com>