Mock with queue (#143)
* replace single register with queue

This not only simplifies the logic.  It is also simpler and more flexible to
use.  For example, multiple responses can be staged in advance without waiting.

* example test with two staged responses

The withdrawal function makes two calls to the ledger if the first
one fails with BadFee. Here we stage both responses at once, then
trigger withdrawal, then release the responses.

* Add Response states #staged, #running, #delivered

* two detailed examples